Move page collection class from the twig plugin to the core so the delay loading works for any templating plugin.
